face-recognition/requirements.txt
opencv-python>=4.8
numpy>=1.24
scipy>=1.10
mtcnn>=0.1.1
keras-facenet>=0.3.2
tensorflow>=2.12
colorama>=0.4.6
google-api-python-client>=2.120
google-auth>=2.30
python-dotenv>=1.0
